Initial ID Page Changes
Changes ID page button when no ID is found to a student ID shaped button with a scan prompt. Changes ID page when an ID is found to a student ID shaped button that has a placeholder image and text in it currently. On tapping this button the user can scan their ID. Added an option to delete user NFC data from Firebase database. Removes ID Testing page.